webkitgtk-2.34.6+wayfarer-3

+++ Source/WebKitLegacy/Storage/StorageTracker.h
@@ -25,6 +25,7 @@
 
 #pragma once
 
+#include <wtf/FastMalloc.h>
 #include <WebCore/SQLiteDatabase.h>
 #include <wtf/HashSet.h>
 #include <wtf/Seconds.h>
@@ -54,7 +55,7 @@ public:
     void deleteOrigin(const WebCore::SecurityOriginData&);
     void deleteOriginWithIdentifier(const String& originIdentifier);
     Vector<WebCore::SecurityOriginData> origins();
-    long long diskUsageForOrigin(WebCore::SecurityOrigin*);
+    uint64_t diskUsageForOrigin(WebCore::SecurityOrigin*);
     
     void cancelDeletingOrigin(const String& originIdentifier);
